Travel Booking Service
The main objective of this Travel & Tourism project is to make travel easy and comfortable for the users right from finding the routes and buses to till booking of the tickets.

Travel and Tour Management Java web project using JDBC, Servlet, and MYSQL using STS tool that is deployed on localhost using tomcat server. This project is using MVC architecture and using Maven tool to manage the dependencies.